gf_lang - more fixes around number casting and more test programs working

+++ gf_lang/go/gf_lang/gf_state_setters.go
@@ -215,7 +215,7 @@ func execExpr(pSetterTypeStr string,
         //------------------------------------
         // ITERS_MAX
 
-        iterationsMaxInt := pVals.(int)
+        iterationsMaxInt := int(pVals.(float64))
         pState.ItersMaxInt = iterationsMaxInt
 
         //------------------------------------
